--- kubemarine/resources/schemas/mount_fs.json
@@ -1,15 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema",
-  "type": "object",
-  "properties": {
-    "reboot": {
-      "type": "boolean",
-      "description": "Whether to drain and reboot the node after mounting. Defaults to true."
-    },
-    "fsmount": {
-      "$ref": "definitions/services/fsmount.json"
-    }
-  },
-  "required": ["fsmount"],
-  "additionalProperties": false
-}
