+++ kubemarine/templates/zram-setup.service.j2
@@ -1,15 +1,13 @@
 [Unit]
-Description=Create zram device
+Description=Create zram mount for path {{ path }}
 DefaultDependencies=no
 Before=local-fs.target
 
 [Service]
 Type=oneshot
 RemainAfterExit=yes
-ExecStart=/usr/sbin/modprobe zram
-ExecStart=/usr/sbin/zramctl {{ device }} --size {{ size }} --algorithm zstd
-ExecStart=/usr/sbin/mkfs.{{ type }} -m 0 -O ^has_journal -F {{ device }}
-ExecStart=/usr/bin/mount {{ device }} {{ path }}
+ExecCondition=/bin/bash -c '! zramctl -o MOUNTPOINT --noheadings | grep -x {{ path }}' 
+ExecStart=/bin/bash -c 'modprobe zram; mkdir -p {{ path }}; device=$(zramctl --find --size {{ size }} --algorithm zstd); mkfs.ext4 -m 0 -O ^has_journal -F $device; mount $device {{ path }}'
 
 [Install]
 WantedBy=local-fs.target
